src/test/java/com/darx/tree/TreesTest.java
package com.darx.tree;

import org.junit.Assert;
import org.junit.Test;

import java.util.ArrayList;
import java.util.List;

import static org.junit.Assert.*;

public class TreesTest {

    @Test
    public void preorderTraversal() {
        Trees trees = new Trees();
        TreeNode root = new TreeNode(1, null, new TreeNode(2, new TreeNode(3), null));

        List<Integer> result = trees.preorderTraversal(root);

        Assert.assertEquals(Integer.valueOf(1), result.get(0));
        Assert.assertEquals(Integer.valueOf(2), result.get(1));
        Assert.assertEquals(Integer.valueOf(3), result.get(2));
    }

    @Test
    public void postorderTraversal() {
        Trees trees = new Trees();
        TreeNode root = new TreeNode(1, null, new TreeNode(2, new TreeNode(3), null));

        List<Integer> result = trees.postorderTraversal(root);

        Assert.assertEquals(Integer.valueOf(3), result.get(0));
        Assert.assertEquals(Integer.valueOf(2), result.get(1));
        Assert.assertEquals(Integer.valueOf(1), result.get(2));
    }
}